requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-multipart
python-dotenv
pypdf
